requirements.txt
anthropic>=0.40.0
rich>=13.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Web GUI (FastAPI)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
